requirements.txt
streamlit==1.41.1
openai==2.33.0
duckdb==1.5.2
pandas==2.3.3
numpy==1.26.4
agno==2.6.2
python-dotenv==1.2.2
